pro-2630 Add lightbox "are you sure" when removing credential

+++ mikoba/ViewModels/Pages/EntryHubPageViewModel.cs
@@ -1,32 +1,22 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using System.Windows.Input;
-using Autofac;
 using DynamicData;
 using Hyperledger.Aries.Agents;
-using Hyperledger.Aries.Configuration;
 using Hyperledger.Aries.Contracts;
 using Hyperledger.Aries.Features.DidExchange;
 using Hyperledger.Aries.Features.IssueCredential;
-using Hyperledger.Aries.Features.PresentProof;
 using Hyperledger.Aries.Routing;
-using Hyperledger.Aries.Storage;
-using mikoba.CoreImplementations;
 using mikoba.Extensions;
 using mikoba.Services;
 using mikoba.UI.Helpers;
 using mikoba.ViewModels.Components;
 using mikoba.ViewModels.SSI;
-using Newtonsoft.Json;
 using ReactiveUI;
-using West.Extensions.Xamarin;
-using Xamarin.Essentials;
 using Xamarin.Forms;
-using CredentialPreviewAttribute = mikoba.ViewModels.Components.CredentialPreviewAttribute;
+using Acr.UserDialogs;
 using INavigationService = mikoba.Services.INavigationService;
 
 namespace mikoba.ViewModels.Pages
@@ -39,8 +29,8 @@ namespace mikoba.ViewModels.Pages
             ICredentialService credentialService,
             IEdgeClientService edgeClientService,
             IAgentProvider contextProvider,
-            IDialogService dialogService,
-            IEventAggregator eventAggregator
+            IEventAggregator eventAggregator,
+            IUserDialogs userDialogs
         )
             : base("Hub Page", navigationService)
         {
@@ -50,17 +40,17 @@ namespace mikoba.ViewModels.Pages
             _credentialService = credentialService;
             _eventAggregator = eventAggregator;
             _edgeClientService = edgeClientService;
-            _dialogService = dialogService;
+            _userDialogs = userDialogs;
         }
 
         #region Services
-        
+
         private readonly IConnectionService _connectionService;
         private readonly ICredentialService _credentialService;
         private readonly IAgentProvider _contextProvider;
         private readonly IEventAggregator _eventAggregator;
         private readonly IEdgeClientService _edgeClientService;
-        private readonly IDialogService _dialogService;
+        private readonly IUserDialogs _userDialogs;
 
         #endregion
 
@@ -85,13 +75,12 @@ namespace mikoba.ViewModels.Pages
         {
             try
             {
-                //TODO: Move away from this library that has non implemented methods
-                // var userConfirmation = await _dialogService.ShowAlertAsync("Are you sure you want to remove this credential?",
-                //     "Confirmation", "Yes", "No, Cancel");
-                // if (!userConfirmation)
-                // {
-                //     return;
-                // }
+                var userConfirmation = await _userDialogs.ConfirmAsync("Are you sure you want to remove this credential?",
+                    "Confirmation", "Yes", "No, Cancel");
+                if (!userConfirmation)
+                {
+                    return;
+                }
                 var context = await _contextProvider.GetContextAsync();
                 await _credentialService.DeleteCredentialAsync(context, _credential._credential.Id);
                 _eventAggregator.Publish(new CoreDispatchedEvent() {Type = DispatchType.ConnectionsUpdated});
